requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argilla==1.5.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
backoff==2.2.1
beautifulsoup4==4.12.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromadb==0.3.20
click==8.1.3
clickhouse-connect==0.5.18
cohere==4.1.3
colorama==0.4.6
commonmark==0.9.1
confection==0.0.4
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
Deprecated==1.2.13
dill==0.3.6
duckdb==0.7.1
effdet==0.3.0
elastic-transport==8.4.0
elasticsearch==8.6.2
et-xmlfile==1.1.0
faiss-cpu==1.7.3
fastapi==0.95.0
filelock==3.10.7
fonttools==4.39.3
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
iopath==0.1.10
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.130
langcodes==3.3.0
layoutparser==0.3.4
lxml==4.9.2
lz4==4.3.2
manifest-ml==0.0.1
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.0
nltk==3.8.1
numpy==1.23.5
olefile==0.46
omegaconf==2.3.0
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
pathy==0.10.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.8.0
Pillow==9.5.0
portalocker==2.7.0
posthog==2.4.2
preshed==3.0.8
pycocotools==2.0.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.14.0
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.7.0
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pywin32==306
PyYAML==6.0
redis==4.5.4
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.47
sqlitedict==2.1.0
srsly==2.4.6
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
thinc==8.1.9
threadpoolctl==3.1.0
tiktoken==0.3.3
timm==0.6.13
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.27.4
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.5.9
unstructured-inference==0.2.4
urllib3==1.26.15
uvicorn==0.21.1
Wand==0.6.11
wasabi==1.1.1
watchfiles==0.19.0
websockets==11.0
wikipedia==1.4.0
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
zstandard==0.20.0
